Added many internal providers to the API for public use (#807)
