+++ API/src/main/java/com/bgsoftware/superiorskyblock/api/hooks/AFKProvider.java
@@ -4,6 +4,11 @@ import org.bukkit.entity.Player;
 
 public interface AFKProvider {
 
+    /**
+     * Check whether a player is considered AFK.
+     *
+     * @param player The player to check.
+     */
     boolean isAFK(Player player);
 
 }
