--- src/main/java/com/bgsoftware/superiorskyblock/hooks/provider/PermissionsProvider.java
@@ -1,9 +1,0 @@
-package com.bgsoftware.superiorskyblock.hooks.provider;
-
-import org.bukkit.entity.Player;
-
-public interface PermissionsProvider {
-
-    boolean hasPermission(Player player, String permission);
-
-}
