+++ src/main/java/com/bgsoftware/superiorskyblock/hooks/provider/PricesProvider_Default.java
@@ -1,6 +1,7 @@
 package com.bgsoftware.superiorskyblock.hooks.provider;
 
-import com.bgsoftware.superiorskyblock.key.Key;
+import com.bgsoftware.superiorskyblock.api.hooks.PricesProvider;
+import com.bgsoftware.superiorskyblock.api.key.Key;
 
 import java.math.BigDecimal;
 
@@ -9,7 +10,7 @@ public final class PricesProvider_Default implements PricesProvider {
     private final BigDecimal INVALID_WORTH = BigDecimal.valueOf(-1);
 
     @Override
-    public BigDecimal getPrice(com.bgsoftware.superiorskyblock.api.key.Key key) {
+    public BigDecimal getPrice(Key key) {
         return INVALID_WORTH;
     }
