--- src/main/java/com/bgsoftware/superiorskyblock/hooks/provider/VanishProvider.java
@@ -1,9 +1,0 @@
-package com.bgsoftware.superiorskyblock.hooks.provider;
-
-import org.bukkit.entity.Player;
-
-public interface VanishProvider {
-
-    boolean isVanished(Player player);
-
-}
